Removed error detection from default I2C code. Fixes #2
Besides requiring the "one additional handshake" the error detection
code suggested in the ATmega328P datasheet was also causing problems
when included in student code. Experimentation in a simulator indicated
that a "data received ACK" was present after a START bit was sent. My
working hypothesis is that the problem happens when the system is reset
while the I2C peripheral is expecting data, causing a misinterpretation
of the START bit.

With only one controller and only one peripheral on the circuit (and
that peripheral having a visible indication of successful operation) the
error-detection code isn't strictly necessary.