Removed bug that constrained the include order for nested headers
When the Arduino IDE autommatically included a nested header file
upon auto-complete, it caused some headers (such as cowpi_io.h)
that depended on stdbool.h to be included before stdbool.h was
included.

The fix is to make sure that the nested headers include any
standard headers they actually depend on, instead of relying on
the inclusion order.