Added file stream for full LCD1602 display
Where the row-based file streams allow the programmer to determine text
placement by using one stream or another, the full-display file stream
requires the programmer to use ASCII placement codes -- but it's
possible that the `\n` behavior is more intuitive by doing so.

It's probable that only one approach or the other will still be around
when we get to the 1.0 release.